Set up TURN server container for Synapse.
Not verified that it works; only verified that Synapse runs with it configured.
I suspect that it still needs a valid exposed UDP port range, and it may also
require having the TURN server detect the external IP of the host before
starting. This may in turn require setting --network=host, as it is done on
https://github.com/bprodoehl/docker-turnserver